Float

+++ teor_ver/Laba_6.py
@@ -2,7 +2,7 @@ mass = []
 with open('C:\\Users\\Alexandr\\Documents\\GitHub\\Python-projects\\teor_ver\\file.txt',"r") as file:
     lines = file.readlines()
     for line in lines: 
-        read_line = [int(x) for x in line.split()]
+        read_line = [float(x) for x in line.split()]
         mass.append(read_line)
     for i in range (len(mass)):
         print ("i = " + str(i))
